requirements.txt
psycopg2
Pillow
Django
gunicorn
django-heroku
whitenoise
djangorestframework
markdown
django-filter